WAIFU-28: refactor: move user business logic to UserService
