+++ src/service.rs
@@ -1,1 +1,2 @@
 pub mod setup;
+pub mod user;
